Added fvSchemes class
Now, fvSchemes can be controlled as well through a dictionary entry
in the main python file. Support is added for various numerical schemes
and orders for the time integration and convective fluxes, as well as
support for various degrees of limitations and corrections to be applied
to various numerical schemes. The current arrangement, however, only
exposes a 4-tier system where users can select an option from 4 pre-
defined settings which will then automatically apply the corrections in
the right places, without the user needing to modify the fvSchemes file
directly.