Refactored initial turbulent quantities calculation
Some quantities were calculated in a way that gave incorrect drag
coefficient values. These have been fixed and now the user can specify
how to calculate the intial quantities. Either based on internal or
external flow, or, if this does not apply, the turbulent quantities can
be based on the turbulent to laminar viscosity ratio (either manual
or automatic calculation, based on the freestream turbulent intensity)