Updated fvScheme correction values
The default fvScheme policy was changed from 1 to 0.33 which seems to
provide enough stability throughout the domain.

This may need to be revisited in the future in case simulations are
diverging and stronger corrections are required. In which case, it may
be beneficial to introduce a correction variable rather than changing
the whole fvScheme discretisation policies.