Added force coefficient convergence checking
Now we are able to check whether the solution has converged based on
integral quantities (force coefficients). A trigger mechanism has been
set up where a minimum number of timesteps need to pass first before
any convergence is checked. After that, a quantity of interest is
monitored and averaged over a period of time steps (specified by
the user). If the average force coefficient within that window does not
change more than a user specified convergence threshold, convergence
is triggered and the simulation stopped.