Refactored properties dictionaries
The number of dictionaries was ever increasing and got a bit difficult to
handle, with several input dictionaries being passed to classes. Now, a single
dictionary called 'properties' exist which includes all sub-dictionaries.
Only this single property object is passed to classes, simplifying the passing
of arguments to classes. This will also allow to easily introduce further
dictionaries in the future without having to modify the file writing classes'
datastructures.

This change was introduced to simplyfy the changes to be introduced for
compressible file writing capabilities and to simplify the current handling
of reference values and force coefficient storage (currently in two different
dictionaries). Similar changes may be needed for the residual handling.