Added 0-D point probe sampling
Now the user can specify to sample 0-D points and write out flow quantities of
interest (such as velocity, pressure and turbulent properties). This may be
particularly useful for scale-resolved simulations to collect data on the
energy spectra, but may also be used for non scale-resolved simulations to
monitor the level of convergence throughout the flow field.