Added iso-surface writing
Now iso-surfaces can be directly exported during calculation, which may be
beneficial for large simulations where iso-surfaces are of interest, rather
than storing the whole 3-D flow field.

Currently this is only working for internal variables, i.e. those calculated
by the solver anyways and not user-defined variables, i.e. those calculated
on-the-fly through a post-processing routine (like vorticity or the
Q-criterion). This needs to be fixed (can be achieved by placing field
calculation and iso surface generation in the same file) but ideally we would
like to keep them in separate files for modularity.