Modified iso-surface writing
The issue of not being able to write iso-surfaces of additional flow variables
has been resolved by forcing the iso-surfaces to be computed after the
additional flow variables (by including its calculation / generation in the
fields files). This may be a sub-optimal solution but it seems there is no
other easy way around.