Added json file writing and mesh copying support
Now the user can specify to write out a case setup as a json file which in
turn can also be read back into the case generator to setup a previously
defined case. To accommodate this, command line parsing has been implemented
which can be shown using the --help command line argument.

In addition, the user can now also specify the location of the mesh to be
used, which, if specified, will be copied into the case setup. This could be
further improved by checking that the specified boundary conditions actually
do exist in the mesh (which has not yet been implemented).

One example case for a NACA 0012 airfoil has been added, including the mesh,
which can be used as an example to see how the case generator can be used.