Allow for general mesh treatment
Now the user can specify if the mesh should be read from a polyMesh dir
as before, or, if the blockMeshDict (and potentially the
snappyHexMeshDict) file should be simply copied to the system dir.
The user can also choose not to do anything with the mesh, meaning that
they require to copy the mesh themselves and a message will be printed
to screen to inform them to do so.

The Allrun script was updated accordingly to allow for blockMesh or
snappyHexMesh to be invoked automatically. The parallel running was
fixed as well as it contained a wrong syntax. The fomaJob command has
been removed as it is no longer required.