Refactored boundary condition writing
The boundary condition class was refactored to more easily add content
(in other words, the open-closed principle was more strictly enforced).

The custom initial and boundary condition writing capabilities were
also completely re-written. Instead of writing a c++ file stub, the
user can now specify a folder in which a custom c++ file is located.
The case generator will now take that file and insert that as the
custom code and generate the OpenFOAM boilerplate code around it
required for dynamic code compilation.