Changed the README.md file
The documentation was completely rewritten to reflect the changes made
to the refactored source code.

In addition, the GlobalVariables.py module was moved into the input
directory to bring it closer to the source where it is used, i.e. in
the CaseProperties.py module. This was a bit of cleanup performed
while writing the documentation as it was deemed to be a better
location for this module.