Added post-processing dictionary
Some cases will require some user-defined post-processing, be it
function objects directly executed within OpenFOAM or python scripts
that run after the simulation is done. The post-processing dictionary
allows for exactly that and acknowledges the complex nature of
post-processing, i.e. various different function objects can be
executed and user-defined post-processing routines written in python
may further enhance automatic data extraction at the end of the
simulation.

This functionality was added to support the post-processing of the
Taylor-Green Vortex problem to calculate the kinetic energy dissipation
rate and the evolution of the kinetic energy.