Some file clean-up
Some clean-up was required in the src directory to adjust some specific
case setup. The Taylor Green Vortex problem is now added and fully
supported, although the results show that potentially an explicit
solver with Runge Kutta time stepping may be a better choice.
Some issues with the pressure coefficient writing were also discovered
which was fixed. The Solution and Schemes files were also further
fine-tuned to work with a larger set of problems based on
experimentation.