Adjusted compressible min/max pressure handling
OpenFOAM allows for specifying either relative or absolute pressure
values that should be respected at boundaries. The relative version
seems to be more robust, yet if no inlet / outlet is provided, there is
no reference pressure level and the absolute min / max values need to be
prescribed. This is handled now automatically.