Refactored case properties handling
Before, a global CaseProperties.py file existed which was responsible
for handling all case input and then some additional default processing
on top of the provided input. This has now been split and an abstract
base clase CasePropertiesBase.py has been introduced, from which
different cases can inherit their properties now. This ensures each
case property file will provide a common interface for the loader
which will dynamically load the right module with the corresponding
case properties, which is specified through a command line argument.

The JSON loader has been removed as a result to retain only a single
truth within the whole project. This also means that if further
properties are added to the case properties dictionary, this change
needs to propagate to all modules within the input directory. However,
this allows to have tailored input script which don't have to be
overwritten once a new case should be set up and it provides a better
interface than the previous JSON solution.

The CaseProperties.py file is now only repsonsible for handling input
and processing it, but it does not take any user-input and thus it has
been moved into the src directory. User input is still handled in the
input directory but now each case has its own python module, which is
specified via the command line.