Refactored snappyHexMesh treatment
The snappyHexMesh treatment was refactored to allow the user more
flexibility in how to steer the meshing stage. Now a snappyHexMeshDict
file is copied to the system location and then, depending on
availability, either a blockMeshDict file or a polyMesh folder is copied
into the appropriate location to provide a background mesh for
snappyHexMesh. Additionally, the user can also specify a list of geometry
files which will be copied into the constant/triSurface folder. In this
way, the pre-processing can be dealt with entirely automatically.

A example of a wing + winglet configuration was added to test the
snappyHexMesh automatic case setup capabilities.