Added system tests
Testing was always an issue, as small unit tests could be provided,
however, these could not cover the entire project and thus are less
helpful. Therefore, now with a matured case generator, system tests
have been put in place that test the automatic case setup. Currently no
effort is made to run these tests in OpenFOAM but this may as well be
added as another testing layer later. However, these system tests allow
to test that different parts of the code still work as expected and
produce a ready case setup.