Refactored codedFixedValue boundary conditions
Before we used a fixedValue boundary condition with a codeStream
statement which was only setting the boundary conditions at the first
time step (e.g. in the 0/ directory). This was refactored to now use
the codedFixedValue boundary condition which is called after each
iteration and allows to control time dependent boundary conditions.