Refactored boundary condition writing
Now the calculation of turbulent freestream parameters is delegated to its
own class that will calculate these parameters in the constructor of the
boundary condition class. As a result, the boundary condition class has been
upgraded to become a management class and its name has changed to reflect
its new responsibilities.