Added new case setup example scripts, WIP
New example scripts were added which currently are still work in progress.
These show, however, how to work with the new module structure instead of
json file reading, which is more intuitive.

Each case comes with it own README file for clarity, so that the usage and
expected results are documented.

Results do currently not necessarily match the reference data and this is
currently worked on.