Refactored boundary condition writing
The state variables are now being handled externally and the boundary
condition manager does delegate that responsibility to the state variable
manager class. This class will determine which variables are required for the
simulations which can then be received by the caller. The boundary condition
manager uses them now to determine which variables need to be written into
the 0 directory.