Refactored fvSolution file writing
Now each state variable is explicitly stated in the fvSolution file and
the linear solver can be adjusted on a per variable basis. This allows
fine tuning but also adds clarity of which variables are solved for
which is more consistent with the refactored case generator, which now
only writes out variables that are actually being used.