Changed time integration policy
Before this commit, in order to change from steady to unsteady
calculations, one had to change the input file in different places
which put extra effort on changing a case from steady to unsteady.
This commit refactored the time-integration policy in that it all
responsibility is now delegated to the time discretisation setting.
If we set the case to be steady or unsteady state, not only will the
time discretisation be changed accordingly, but all required changes
will be propagated down the line, i.e. the controlDict will be set up
accordingly (which also means that we have now two sets of input, one
for steady and one for unsteady flows). These new inputs are stored
within the file_properties/time_discretisation map. Accordingly, the
entries in file_properties/solver_properties have changed (no more
controlDict setup is done here) and also it has gained access over the
correctors (nCorrectors, nOuterCorrectors, nNonOrthoCorrectors).
Since the numerical discretisation has lost its time derivative
setting, it has been renamed to spatial discretisation to reflect that
only spatial terms are set up here.