Added unit and system tests
The system test was removed and repalced by a more user friendly and
extendible test runner. Furthermore, all file output writing is now
decoupled from the classes that generate the string to be written to
file; the filemanager class writes the content emitted by string
producing classes, which further allowed to introduce unit testing more
easily.

As a result, some refactoring was done. The Global Properties variable
is gone and was replaced by Enum classes. The main file deligates now
the case generation to its own class and only manages the flow of
information. A few adjustments were made to make the code more concise,
as well as some bugs were fixed that came out as a result of the more
rigorous testing compared to before.